Update and implement cors global, spring version, some apis about appointment, conversation, account

+++ src/main/java/com/gsu21se45/entity/AveragePrice.java
@@ -37,10 +37,10 @@ public class AveragePrice implements Serializable {
     private double price;
 
     @Column(name = "month")
-    private double month;
+    private int month;
 
     @Column(name = "year")
-    private double year;
+    private int year;
 
     @Column(name = "type")
     private String type;
